Update to Smithay 0.30.0
This patch is a big rewrite of most of the compositor interface to
Smithay's new trait-based API. In the process some other minor bugs were
also resolved.

Layer shell exclusive zones were not properly reset, since their
reservation was re-applied right after recovery due to the
`update_dimensions` call performed on each layer window when the
exclusive zone changed. To fix this, the exclusive zone update has been
extracted to only be performed on surface commit.

Some UI elements like border/decorations was incorrectly using the DPI
scaling factor to always render at a fixed size, which defeats the
purpose of having a DPI scaling factor. This has been adjusted to ensure
bigger factor means bigger UI.

After updating to the latest Smithay master and rewriting the virtual
keyboard patch, the issue with breaking touch once a virtual
keyboard/IME was started has also been resolved.

+++ src/layer.rs
@@ -49,7 +49,7 @@ impl Layers {
         renderer: &mut Gles2Renderer,
         frame: &mut Gles2Frame,
         output: &Output,
-        damage: &[Rectangle<f64, Physical>],
+        damage: &[Rectangle<i32, Physical>],
     ) {
         for window in &mut self.background {
             window.draw(renderer, frame, output, 1., None, damage);
@@ -66,7 +66,7 @@ impl Layers {
         renderer: &mut Gles2Renderer,
         frame: &mut Gles2Frame,
         output: &Output,
-        damage: &[Rectangle<f64, Physical>],
+        damage: &[Rectangle<i32, Physical>],
     ) {
         for window in &mut self.top {
             window.draw(renderer, frame, output, 1., None, damage);
